+++ src/js/components/filter.js
@@ -1,7 +1,10 @@
 import Animate from '../mixin/animate';
 import {
+    $,
     $$,
     append,
+    attr,
+    closest,
     css,
     data,
     each,
@@ -11,6 +14,7 @@ import {
     includes,
     isEmpty,
     isEqual,
+    isTag,
     isUndefined,
     matches,
     parseOptions,
@@ -42,12 +46,18 @@ export default {
                 return $$(`[${attrItem}],[data-${attrItem}]`, $el);
             },
 
-            watch() {
+            watch(toggles) {
                 this.updateState();
 
-                if (this.selActive !== false) {
-                    const actives = $$(this.selActive, this.$el);
-                    this.toggles.forEach((el) => toggleClass(el, this.cls, includes(actives, el)));
+                const actives = $$(this.selActive, this.$el);
+                for (const toggle of toggles) {
+                    if (this.selActive !== false) {
+                        toggleClass(toggle, this.cls, includes(actives, toggle));
+                    }
+                    const button = findButton(toggle);
+                    if (isTag(button, 'a')) {
+                        attr(button, 'role', 'button');
+                    }
                 }
             },
 
@@ -78,8 +88,10 @@ export default {
             },
 
             handler(e) {
-                e.preventDefault();
-                this.apply(e.current);
+                if (closest(e.target, 'a,button')) {
+                    e.preventDefault();
+                    this.apply(e.current);
+                }
             },
         },
     ],
@@ -108,9 +120,11 @@ export default {
 
             trigger(this.$el, 'beforeFilter', [this, state]);
 
-            this.toggles.forEach((el) =>
-                toggleClass(el, this.cls, !!matchFilter(el, this.attrItem, state))
-            );
+            for (const toggle of this.toggles) {
+                const active = !!matchFilter(toggle, this.attrItem, state);
+                toggleClass(toggle, this.cls, active);
+                attr(findButton(toggle), 'aria-selected', active);
+            }
 
             await Promise.all(
                 $$(this.target, this.$el).map((target) => {
@@ -210,4 +224,8 @@ function sortItems(nodes, sort, order) {
             data(a, sort).localeCompare(data(b, sort), undefined, { numeric: true }) *
             (order === 'asc' || -1)
     );
+}
+
+function findButton(el) {
+    return $('a,button', el) || el;
 }