+++ src/js/components/tooltip.js
@@ -20,6 +20,7 @@ import {
     remove,
     within,
 } from 'uikit-util';
+import { generateId } from '../mixin/utils';
 
 export default {
     mixins: [Container, Togglable, Position],
@@ -41,7 +42,7 @@ export default {
     },
 
     beforeConnect() {
-        this.id = `uk-tooltip-${this._uid}`;
+        this.id = generateId(this);
         this._hasTitle = hasAttr(this.$el, 'title');
         attr(this.$el, {
             title: '',