+++ src/js/core/icon.js
@@ -1,4 +1,5 @@
 import SVG from './svg';
+import I18n from '../mixin/i18n';
 import closeIcon from '../../images/components/close-icon.svg';
 import closeLarge from '../../images/components/close-large.svg';
 import marker from '../../images/components/marker.svg';
@@ -23,13 +24,16 @@ import {
     $,
     addClass,
     apply,
+    attr,
     closest,
     css,
     each,
+    hasAttr,
     hasClass,
     hyphenate,
     isRtl,
     isString,
+    isTag,
     parents,
     swap,
 } from 'uikit-util';
@@ -135,16 +139,12 @@ export const Search = {
     },
 };
 
-export const Close = {
+export const Spinner = {
     extends: IconComponent,
 
     beforeConnect() {
-        this.icon = `close-${hasClass(this.$el, 'uk-close-large') ? 'large' : 'icon'}`;
+        attr(this.$el, 'role', 'status');
     },
-};
-
-export const Spinner = {
-    extends: IconComponent,
 
     methods: {
         async getSvg() {
@@ -157,6 +157,60 @@ export const Spinner = {
             return icon;
         },
     },
+};
+
+const ButtonComponent = {
+    extends: IconComponent,
+
+    mixins: [I18n],
+
+    beforeConnect() {
+        const button = closest(this.$el, 'a,button');
+
+        attr(button, 'role', this.role !== null && isTag(button, 'a') ? 'button' : this.role);
+
+        const label = this.t('label');
+        if (label && !hasAttr(button, 'aria-label')) {
+            attr(button, 'aria-label', label);
+        }
+    },
+};
+
+export const NavbarToggleIcon = {
+    extends: ButtonComponent,
+    i18n: { label: 'Open menu' },
+};
+
+export const Close = {
+    extends: ButtonComponent,
+
+    i18n: { label: 'Close' },
+
+    beforeConnect() {
+        this.icon = `close-${hasClass(this.$el, 'uk-close-large') ? 'large' : 'icon'}`;
+    },
+};
+
+export const Marker = {
+    extends: ButtonComponent,
+    i18n: { label: 'Open' },
+};
+
+export const Totop = {
+    extends: ButtonComponent,
+    i18n: { label: 'Back to top' },
+};
+
+export const PaginationNext = {
+    extends: ButtonComponent,
+    i18n: { label: 'Next page' },
+    data: { role: null },
+};
+
+export const PaginationPrevious = {
+    extends: ButtonComponent,
+    i18n: { label: 'Previous page' },
+    data: { role: null },
 };
 
 const parsed = {};