+++ src/js/mixin/slider.js
@@ -1,7 +1,8 @@
+import I18n from './i18n';
 import Resize from './resize';
-import SliderAutoplay from './slider-autoplay';
-import SliderDrag from './slider-drag';
 import SliderNav from './slider-nav';
+import SliderDrag from './slider-drag';
+import SliderAutoplay from './slider-autoplay';
 import {
     $,
     children,
@@ -15,7 +16,7 @@ import {
 } from 'uikit-util';
 
 export default {
-    mixins: [SliderAutoplay, SliderDrag, SliderNav, Resize],
+    mixins: [SliderAutoplay, SliderDrag, SliderNav, Resize, I18n],
 
     props: {
         clsActivated: Boolean,