+++ src/js/util/lang.js
@@ -119,7 +119,7 @@ export function toNode(element) {
 }
 
 export function toNodes(element) {
-    return (element && (isNode(element) ? [element] : Array.from(element).filter(isNode))) || [];
+    return isNode(element) ? [element] : Array.from(element || []).filter(isNode);
 }
 
 export function toWindow(element) {